+++ include/ssm-proxy-client.hpp
@@ -27,11 +27,11 @@ private:
 	const char *streamName;
 	int streamId;
 	void *mData;								///< データのポインタ
-	size_t mDataSize;						///< データ構造体のサイズ
+	uint64_t mDataSize;						///< データ構造体のサイズ
 	void *mProperty;							///< プロパティのポインタ
-	size_t mPropertySize;					///< プロパティサイズ
+	uint64_t mPropertySize;					///< プロパティサイズ
 	void *mFullData;
-	size_t mFullDataSize;
+	uint64_t mFullDataSize;
 	ssmtime *timecontrol;					///< for get real time
 	char *ipaddr;
 
@@ -46,11 +46,11 @@ private:
 	void readRawData(char **p, char *d, int len);
 
 	void serializeMessage(ssm_msg *msg, char *buf);
-	bool createRemoteSSM( const char *name, int stream_id, size_t ssm_size, ssmTimeT life, ssmTimeT cycle );
-	bool setPropertyRemoteSSM(const char *name, int sensor_uid, const void *adata, size_t size);
+	bool createRemoteSSM( const char *name, int stream_id, uint64_t ssm_size, ssmTimeT life, ssmTimeT cycle );
+	bool setPropertyRemoteSSM(const char *name, int sensor_uid, const void *adata, uint64_t size);
 	bool getPropertyRemoteSSM(const char *name, int sensor_uid, const void *adata);
 
-	bool sendData(const char *data, size_t size);
+	bool sendData(const char *data, uint64_t size);
 
 public:
 	SSM_tid timeId; // データのTimeID (SSM_tid == int)
@@ -75,7 +75,7 @@ public:
 	bool initRemote();
 	bool initSSM();
 	void setStream(const char *streamName, int streamId);
-	void setBuffer(void *data, size_t dataSize, void *property, size_t propertySize, void *fulldata);
+	void setBuffer(void *data, uint64_t dataSize, void *property, uint64_t propertySize, void *fulldata);
 	void setIpAddress(char *address);
 	bool create(const char *streamName, int streamId, double saveTime, double cycle);
 	bool create(double saveTime, double cycle);
